+++ src/app/about/page.tsx
@@ -1,5 +1,30 @@
+import ContactLinks from "@/components/common/ContactLinks";
+import Image from "next/image";
+
 const About = () => {
-  return <div>This is the about page</div>;
+  return (
+    <div className="flex gap-10">
+      <div className="flex w-fit flex-col items-center gap-5">
+        <div className="relative aspect-square w-[200px] overflow-hidden rounded-full">
+          <Image
+            src="/images/blog_profile.jpeg"
+            alt="Author profile image"
+            fill
+            objectFit="cover"
+          />
+        </div>
+
+        <div className="flex flex-col items-center">
+          <p className="mb-1 text-xl font-bold">sqsung</p>
+          <p className="text-t-subtle">Software Engineer</p>
+          <p className="text-t-subtle">Seoul, Korea</p>
+        </div>
+
+        <ContactLinks />
+      </div>
+      <div className="debug-border flex-grow"></div>
+    </div>
+  );
 };
 
 export default About;
